requirements.txt
pymongo
pyyaml
scrapy
